Update extractor.
In this commit the buffer in extractor was removed. Instead, the extractor now accept a @target param
where the caller is responsible to store the desired feature to store into.

The caller could benefit from this change since the extractor now does not store the data state from
the history.