Fix bugs.
Add only_summaries for faster packet handling. Note that to enable
only_summaries correctly, one has to modify the source code of
PyShark. The issue is fixed in

issue:https://github.com/KimiNewt/pyshark/issues/648

but it seems that the fix has not been merged, thus requires a manual
fix.