Update keylog storage in capture.
This is only an update in dry-run mode demo.
Collecting all SSL keys in a single keylog file
will make it too large for Wireshark to perform
decryption. Therefore, we generate keylog.txt
per website.

Legacy capture SHOULD be decrypted using the
keylog.txt in the capture root. The detailed
.pcapng files are listed in README.md for reference.