Add statistics.py
This script is used to analyze the byte counts along with packet counts
for the given protocol of .pcap files in a directory. The purpose is to
provide insights into the stability of each protocol.

The default combinations is HTTP/2 (TCP/TLS/HTTP/2), currently, only
HTTP/2 and HTTP/3 (UDP/QUIC/HTTP/3) are supported.